==> one-way-dkl/all_states_block/6_mutual_info_plots.Rmd <==
---
title: "5_plots"
author: "Ivanna Rodríguez"
date: "`r Sys.Date()`"
output: html_document
---

This is better broken down by state, I think. Like having a markdown file for each state with visualizations and plotting functions you can just call. Also, would it be worth saving individual state shapefiles for ease of loading the shapefile?

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# install.packages(pacman)
pacman::p_load(sf, tidyverse, tidycensus, readxl, magrittr, here, arrow)
```

```{r}
folder_10 <- 'E:/acs5_block_2010_data/'
folder_15 <- 'E:/acs5_block_2015_data/'
folder_20 <- 'E:/acs5_block_2020_data/'

plots_folder <- 'plots/'
```


```{r}
# read in mutual info datasets and filter for top cities to make charts
states <- c("CA", "IL", "NY", "TX", "AZ", "TN" ,"MA", "MI", "GA")
cities <- c("New York", "Los Angeles", "Chicago", "Houston", "Austin", 
            "Phoenix", "Boston", "Memphis", "Detroit", "Atlanta")

# filter datasets to include only cities of interest
get_fltr_data <- function(var) {
  
  df_15 <- read_parquet(paste0(folder_15, paste0("mi_", var, "_15_blkgrp_all_states.parquet"))) %>%
    ungroup()
  df_20 <- read_parquet(paste0(folder_20, paste0("mi_", var, "_20_blkgrp_all_states.parquet"))) %>%
    ungroup()
  
  df_flt_15 <- df_15 %>%
  filter(state_codes %in% states,
         str_detect(cbsa_title, paste(cities, collapse = "|"))) %>%
    mutate(year = 2015)
  
  df_flt_20 <- df_20 %>%
    filter(state_codes %in% states,
         str_detect(cbsa_title, paste(cities, collapse = "|"))) %>%
    mutate(year = 2020)
  
  # 2010 data currently only available for income
  if (var == "income") {
    df_10 <- read_parquet(paste0(folder_10, paste0("mi_", var, "_10_blkgrp_all_states.parquet"))) %>%
      ungroup()
    
    df_flt_10 <- df_10 %>%
      filter(state_codes %in% states,
             str_detect(cbsa_title, paste(cities, collapse = "|"))) %>%
      mutate(year = 2010)
    
    return(rbind(df_flt_10, df_flt_15, df_flt_20))
    }
  
  return(rbind(df_flt_15, df_flt_20))
}
```

```{r}
race_mi <- get_fltr_data("race")
income_mi <- get_fltr_data("income")
educ_mi <- get_fltr_data("educ")
empl_mi <- get_fltr_data("empl")
```


# Mutual Info Bar Graphs

```{r}
glimpse(income_mi)
```

```{r}
my_theme <- function() {
  theme_minimal() +
    theme(
      plot.title = element_text(size = 16, face = "bold"),
      plot.subtitle = element_text(size = 14),
      axis.title = element_text(size = 12, face = "bold"),
      axis.text = element_text(size = 10),
      legend.title = element_text(size = 10, face = "bold"),
      legend.text = element_text(size = 8),
      legend.position = "top",
      legend.justification = "center"
    )
}
```


```{r}
plot_bars <- function(df, var) {
  df$cbsa_title <- factor(df$cbsa_title, levels = unique(df$cbsa_title[order(df$cbsa_title)]))
  
  df %>%
    ggplot(aes(x = avg_dkl_cbsa_pop, y = cbsa_title, 
               fill = as.factor(year))) +
    geom_bar(stat = "identity", position = "dodge", 
             color = "white") + # Removed black border
    geom_text(aes(label = round(avg_dkl_cbsa_pop, 2)), 
              position = position_dodge(width = 0.9), hjust = -0.1, size = 3) +
    labs(title = paste0("Mutual Information by ", var," Bin and Year"),
         subtitle = "for each major CBSA area",
         x = "Mutual Information",
         y = NULL,
         fill = "Year") +
    scale_fill_brewer(palette = "Accent") + 
    scale_x_continuous(labels = function(x) sprintf("%.2f", x)) + 
    scale_y_discrete(labels = function(x) str_wrap(x, width = 25)) +
    my_theme()
}
```


```{r out.height=10 out.width=10}
plot_bars(race_mi, "Race")
ggsave(paste0(plots_folder, "race_mi.png"), width = 10, height = 6, dpi = 300)

plot_bars(income_mi, "Income")
ggsave(paste0(plots_folder, "income_mi.png"), width = 10, height = 6, dpi = 300)

plot_bars(educ_mi, "Education")
ggsave(paste0(plots_folder, "educ_mi.png"), width = 10, height = 6, dpi = 300)

plot_bars(empl_mi, "Employment")
ggsave(paste0(plots_folder, "empl_mi.png"), width = 10, height = 6, dpi = 300)
```
# Mutual Info Boxplot for ALL CBSAs

```{r}
df_15 <- read_parquet(paste0(folder_15, "mi_income_15_blkgrp_all_states.parquet")) %>%
    ungroup() %>% mutate(year = 2015)
df_20 <- read_parquet(paste0(folder_20, "mi_income_20_blkgrp_all_states.parquet")) %>%
    ungroup() %>% mutate(year = 2020)
df_10 <- read_parquet(paste0(folder_10, "mi_income_10_blkgrp_all_states.parquet")) %>%
      ungroup() %>% mutate(year = 2010)
all_income_mi <- rbind(df_10, df_15, df_20)
rm(df_15, df_20, df_10)
```


```{r}
all_income_mi %>%
  ggplot(aes(x = factor(year), y = avg_dkl_cbsa_pop, fill = as.factor(year))) +
  geom_violin(width=0.6, alpha = 0.5, color = NA) +
  geom_boxplot(width=0.4, alpha = 0.7) +
  scale_fill_brewer(palette = "Accent") +
  scale_y_continuous(limits = c(0, 1.4), breaks = seq(0, 2, 0.2)) +
  labs(title = "Average Mutual Information",
       subtitle = "Across All CBSA Areas by Year",
       y = "Average Mutual Information", 
       fill = "Year") +
  my_theme() +
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        axis.text.x = element_text(face = "bold", size = 12)) +
  guides(fill = FALSE)

ggsave(paste0(plots_folder, "mi_boxplot.png"), width = 6, height = 6, dpi = 300)
```


```{r}
# the other way with std dv and mean
avg_mi <- income_mi %>% 
  group_by(year) %>% 
  summarise(sd_mi = sd(avg_dkl_cbsa_pop), 
            avg_mi = mean(avg_dkl_cbsa_pop))

avg_mi %>%
  ggplot(aes(x = factor(year), y = avg_mi, group = factor(year))) + 
  geom_point() +
  geom_errorbar(aes(ymin = avg_mi - sd_mi, 
                    ymax = avg_mi + sd_mi), 
                width = 0.2) +
  labs(title = "Average Mutual Information",
       subtitle = "Across All CBSA Areas by Year, with Standard Deviation") +
  my_theme() +
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank())
```


# Maps


```{r}
# caution: heavy shapefile takes a few mins to load
folder <- 'E:/shapefiles/'
shp_15_tst <- st_read(paste0(folder, "usa_block_groups_2015.geojson"))
```

```{r}
```

```{r}
# NY
```
